% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.double.ipv4}
\alias{as.double.ipv4}
\title{Convert an IPv4 object to an numeric vector}
\usage{
\method{as.double}{ipv4}(x, ...)
}
\arguments{
\item{x}{IPv4 vector}

\item{...}{unused}
}
\description{
Convert an IPv4 object to an numeric vector
}
\examples{
as.numeric(as.ipv4(iptools::ip_random(10)))
}
