% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ipv4.R
\name{as.ipv4}
\alias{as.ipv4}
\title{Convert a character vector of IP(v4) addresses into an \code{ipv4} object}
\usage{
as.ipv4(x)
}
\arguments{
\item{x}{a character vector of IPv4 addresses}
}
\description{
Convert a character vector of IP(v4) addresses into an \code{ipv4} object
}
