\name{inet-datasets}
\title{Datasets included in inet package}

\alias{ptsd_data}

\description{
The package includes a dataset with measurements of 17 PTSD symptoms taken from 344 individuals. See McNally et al. (2015) for more details.
}

\author{Jonas Haslbeck}

\references{

McNally, R. J., Robinaugh, D. J., Wu, G. W., Wang, L., Deserno, M. K., & Borsboom, D. (2015). Mental disorders as causal systems: A network approach to posttraumatic stress disorder. Clinical Psychological Science, 3(6), 836-849.

}

