% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teredo.R
\name{decode_toredo}
\alias{decode_toredo}
\title{Decode Toredo IPv6 strings into components}
\usage{
decode_toredo(x)
}
\arguments{
\item{x}{a character vector of Toredo IPv6 addresses}
}
\value{
a data frame of decoded addresses
}
\description{
Decode Toredo IPv6 strings into components
}
\note{
only valid IPv6 decoded information is retured.
}
\examples{
decode_toredo("2001:0000:4136:e378:8000:63bf:3fff:fdd2")
}
\references{
\url{https://tools.ietf.org/html/rfc4380}
}
