% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-addr-arpa.R
\name{in_addr_arpa}
\alias{in_addr_arpa}
\alias{in_addr_arpa.ipv4}
\alias{in_addr_arpa.character}
\title{Convert IPv4 addresses into in-addr.arpa format}
\usage{
in_addr_arpa(x, ...)

\method{in_addr_arpa}{ipv4}(x, ...)

\method{in_addr_arpa}{character}(x, ...)
}
\arguments{
\item{x}{a vector of IPv4 addresses}

\item{...}{unused}
}
\description{
Convert IPv4 addresses into in-addr.arpa format
}
\examples{
in_addr_arpa(iptools::ip_random(10))
}
\references{
\href{https://tools.ietf.org/html/rfc2317}{RFC 2317}
}
