% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inet-package.R
\docType{package}
\name{inet}
\alias{inet}
\alias{inet-package}
\title{Higher-order Internet Protocol ('IP') Address Methods}
\description{
A collection of tools and methods that expand upon the functionality
of the 'iptools' package.
}
\author{
Bob Rudis (bob@rud.is)
}
