\name{inet-internal}
\title{Internal inet functions}

\alias{input_checks}


\description{Internal inet functions.}

\author{Jonas Haslbeck}

\value{The only internal function is one that performs input checks for the estimation functions. They return informative errors, if the inputs are not specified properly.}

\details{

These are internal functions.

}
