% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octets.R
\name{octets}
\alias{octets}
\title{Turn an IPv4 object or a character vector of IPv4 addresses into a data frame of octets}
\usage{
octets(x)
}
\arguments{
\item{x}{a character vector of IPv4 addresses}
}
\description{
Turn an IPv4 object or a character vector of IPv4 addresses into a data frame of octets
}
\examples{
octets(iptools::ip_random(10))
}
