% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whois.R
\name{whois_arin}
\alias{whois_arin}
\alias{whois_arin.ipv4}
\alias{whois_arin.character}
\title{Lookup WHOIS IP information via the ARIN REST API}
\usage{
whois_arin(x, ...)

\method{whois_arin}{ipv4}(x, ...)

\method{whois_arin}{character}(x, ...)
}
\arguments{
\item{x}{a vector of IPv4 addresses}

\item{...}{unused}
}
\value{
a (usually) large, heavily-nested list structure classed as "arin_whois"
}
\description{
Lookup WHOIS IP information via the ARIN REST API
}
\examples{
whois_arin(iptools::ip_random(10))
}
\references{
\href{https://www.arin.net/resources/whoisrws/whois_api.html}{ARIN Whois-RWS API Documentation};
\href{https://www.arin.net/resources/whoisrws/index.html}{Whois-RWS Info and FAQ}
}
